Update OddsEvens.java
So if I ever get around to it, I would like to decipher why this compiles but crashes at runtime.  I am trying to create some error handling where if a player enters a string char when asked for a number, it simply says, "that is not a number", and to, "please enter a number"

I am using the Scanner.hasNextInt() method to do the checking but I am not sure where it goes in my code and the syntax to use to get it to work.  So I'll just keep trying diff stuff until it works!